+++ .prettierrc.yml
@@ -2,4 +2,3 @@ trailingComma: es5
 singleQuote: true
 printWidth: 120
 tabWidth: 2
-# bracketSpacing: false
