+++ src/app/components/inspector/TokenNodes.tsx
